Fix Database Schema and Entity Class

+++ src/main/java/pl/reverseAuctions/model/auction/AuctionDao.java
@@ -26,7 +26,7 @@ public class AuctionDao implements Entity<Auction> {
     }
 
     @Override
-    public Auction getById(int id) {
+    public Auction getById(Long id) {
         return null;
     }
 }
